requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
typing_extensions==4.12.2
yarl==1.18.0
